Messages: Payload identifiers and Edge: Registrar
Edge Registrar of Edge Message Handlers is in progress.
Added ability to examine a potential Handler method and determine what the
payload type of that message should be.

Note that there may be multiple handler methods for a given payload type,
and that the order that these methods are called will not be guaranteed.